requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pillow>=8.0.0
albumentations>=1.0.0
scikit-image>=0.19.0
requests>=2.28.0
tqdm>=4.65.0
zipfile36>=0.1.3
seaborn>=0.11.0